+++ src/main/java/de/theidler/create_mobile_packages/CreateMobilePackages.java
@@ -18,8 +18,10 @@ import org.slf4j.Logger;
 public class CreateMobilePackages
 {
     public static final String MODID = "create_mobile_packages";
+    public static final String NAME = "Create: Mobile Packages";
     public static final Logger LOGGER = LogUtils.getLogger();
     public static final CreateRegistrate REGISTRATE = CreateRegistrate.create(CreateMobilePackages.MODID);
+    public static final RoboManager ROBO_MANAGER = new RoboManager();
 
     public CreateMobilePackages() {
         onCtor();