+++ src/main/java/de/theidler/create_mobile_packages/entities/robo_entity/states/FlyToTargetState.java
@@ -21,12 +21,11 @@ public class FlyToTargetState implements RoboEntityState {
         BlockPos targetPos = re.getTargetPosition();
         if (targetPos == null)
             return;
-
         if (re.position().distanceTo(targetPos.getCenter()) <= CMPConfigs.server().beeSpeed.get() / 12.0) {
-            if (re.getTargetBlockEntity() != null) {
-                re.setState(new LandingPrepareState());
-            } else if (re.getTargetPlayer() != null) {
+            if (re.getTargetPlayer() != null) {
                 re.setState(new InteractWithPlayerState());
+            } else if (re.getTargetBlockEntity() != null) {
+                re.setState(new LandingPrepareState());
             }
             re.setTargetVelocity(Vec3.ZERO);
         } else {