+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 
 ------------------------------------------------------
-Create: Mobile Packages Unreleased
+Create: Mobile Packages - v0.5.0 - 25.02.2025
 ------------------------------------------------------
 
 ### Changes
@@ -25,6 +25,10 @@ Create: Mobile Packages Unreleased
 - Added a Config option to disable Bee nametags
 - Added Categories to the Portable Stock Ticker (added by linking to a Stock Ticker with Categories)
 - Added rigging to a package carried by a Robo Bee
+- Added when spawning a Robo Bee with the Robo Bee Item while holding a package in the offhand, the package will be rigged to the Robo Bee allowing Player to Player and Player to Port transfer
+- Added Keybind to open a Portable Stock Ticker being in the player inventory (default: `G`)
+- Updated Ponder for the Robo Bee Port
+- Added Tooltip to the Robo Bee Item
 
 ### Bug Fixes
 
@@ -34,6 +38,10 @@ Create: Mobile Packages Unreleased
 - Fixed Portable Stock Ticker address not being saved to the NBT
 - Fixed applied Foil effect to a linked Portable Stock Ticker
 - Changed Packages are no longer their own entity but are now a part of the Robo Bee entity
+- Fixed Robo Bees can't fly through unloaded chunks, by adding a global RoboManager that keeps track of all Robo Bees and ticks them
+- Fixed Offhand block is placed when opening a Portable Stock Ticker
+- Fixed Portable Stock Ticker not working with Create: Factory Logistics
+- Fixed Portable Stock Ticker showing Fluids from Create: Factory Logistics as Air by not showing them at all
 
 ### Translations
 
@@ -41,4 +49,6 @@ Create: Mobile Packages Unreleased
 - Added Chinese (Simplified Han script) translation
 - Added Russian translation
 - Added French translation
+- Added Japanese translation
+- Added Portuguese (Brazilian) translation